+++ modules/School_Setup/Rollover.php
@@ -556,7 +556,7 @@ function Rollover( $table, $mode = 'delete' )
 
 			if ( $mode === 'delete' )
 			{
-				// Fix SQL error foreign key exists on tables GRADEBOOK_ASSIGNMENTS,GRADEBOOK_ASSIGNMENT_TYPES,SCHEDULE_REQUESTS
+				// Fix SQL error foreign key exists on tables GRADEBOOK_ASSIGNMENTS,gradebook_assignment_types,SCHEDULE_REQUESTS
 				// Error happens when an Assignment,or a Schedule request
 				// was added for a rolled-over Course.
 				$delete_sql = "DELETE FROM GRADEBOOK_ASSIGNMENTS
@@ -567,7 +567,7 @@ function Rollover( $table, $mode = 'delete' )
 						WHERE SYEAR='" . $next_syear . "'
 						AND SCHOOL_ID='" . UserSchool() . "');";
 
-				$delete_sql .= "DELETE FROM GRADEBOOK_ASSIGNMENT_TYPES
+				$delete_sql .= "DELETE FROM gradebook_assignment_types
 					WHERE COURSE_ID IN(SELECT COURSE_ID FROM courses
 						WHERE SYEAR='" . $next_syear . "'
 						AND SCHOOL_ID='" . UserSchool() . "');";